Move `message_type::Error` to `error::MessageTypeError`
