+++ src/error.rs
@@ -26,3 +26,28 @@ impl std::fmt::Display for ChannelTypeError {
     }
 }
 
+#[derive(Error, Eq, PartialEq, Clone, Debug)]
+pub enum MessageTypeError {
+    // Marshal buffer was too short
+    UnexpectedEndOfBuffer { expected: usize, actual: usize },
+
+    // DataChannel message has a type we don't support
+    InvalidMessageType { invalid_type: u8 },
+}
+
+impl std::fmt::Display for MessageTypeError {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            Self::UnexpectedEndOfBuffer { expected, actual } => {
+                writeln!(
+                    f,
+                    "Marshal buffer was too short: (expected: {:?}, actual: {:?})",
+                    expected, actual
+                )
+            }
+            Self::InvalidMessageType { invalid_type } => {
+                writeln!(f, "Invalid message type: {:?}", invalid_type)
+            }
+        }
+    }
+}
